Use molo.core version 3.11.1
This latest release of molo.core made changes to how the demotion
of Topic of the day articles are handled. Upon demotion, the Topic
of the day article becomes the most recent article in the Latest
listing. Since Topic of the day articles' promotion can be
future-dated, it is then necessary to omit those future-dated
articles from the Latest articles listing.